server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
bitarray==2.9.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
importlib_resources==6.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.10.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
pycryptodome==3.20.0
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.18.1
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
web3==6.20.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
